fix(nats-manager): fix some errors, and refactor types
