+++ engine/nats-manager/cmd/main.go
@@ -4,16 +4,15 @@ import (
 	"log"
 	"net"
 
-	"github.com/konstellation-io/kre/engine/nats-manager/internal/config"
-	"github.com/konstellation-io/kre/engine/nats-manager/internal/manager"
-
-	"github.com/konstellation-io/kre/engine/nats-manager/nats"
-	"github.com/konstellation-io/kre/engine/nats-manager/proto/natspb"
-	"github.com/konstellation-io/kre/engine/nats-manager/service"
-
 	"github.com/konstellation-io/kre/libs/simplelogger"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/reflection"
+
+	"github.com/konstellation-io/kre/engine/nats-manager/internal/config"
+	"github.com/konstellation-io/kre/engine/nats-manager/internal/manager"
+	"github.com/konstellation-io/kre/engine/nats-manager/internal/service"
+	"github.com/konstellation-io/kre/engine/nats-manager/nats"
+	"github.com/konstellation-io/kre/engine/nats-manager/proto/natspb"
 )
 
 func main() {
