+++ engine/nats-manager/internal/logger/interface.go
